Create README.md
PDF File Search Interface  This Python code provides a graphical user interface (GUI) for searching PDF files in a specified folder. It allows you to enter a folder path and a keyword, and then it searches the PDF files in the folder for occurrences of the keyword. The found files are displayed in a listbox.  Features: - Browse button to select the folder path using a file dialog - Search button to initiate the search process - Listbox to display the found files - Error handling for invalid inputs and folder paths  Dependencies: - Python 3.x - tkinter module for the GUI - PyPDF2 library for reading and extracting text from PDF files  Usage: 1. Run the script. 2. Enter the folder path where the PDF files are located. 3. Enter the keyword to search for within the PDF files. 4. Click the Search button to start the search. 5. The found files will be listed in the interface.  Note: Make sure you have the tkinter and PyPDF2 libraries installed before running the code.  Contributing: - If you find any issues or have suggestions for improvements, please feel free to open an issue or create a pull request.  License: